TaskDirectoryActions cleanup; move fetchNodeState to MesosStateActions
